// angular-sample/angular-webapp/client/app/employee/delegation/models/delegation.ts
export class Delegation {
    UserId: string;
    FirstName: string;
    LastName: string;
    FullName:string;
    Email: string;
    UserFirstName:string;
    UserLastName:string;
    DeligatedUserId: string;
    DeligatedReadOnlyDE: boolean;
    DeligatedManageDE: boolean;
    DeligatedHA: boolean;
    CreatedOn: Date;
    CreatedBy: string;
    Id: string;
}